+++ clj/src/limabean/adapter/loader.clj
@@ -6,7 +6,9 @@
             [clojure.java.io :as io]
             [clojure.string :as str]
             [limabean.adapter.debug :as debug]
-            [limabean.core.type :as type]))
+            [limabean.core.type :as type]
+            [limabean.adapter.exception :as exception]
+            [limabean.adapter.synthetic-spans :as synthetic-spans]))
 
 (defn- dump
   "Dump the beans and return the map"
@@ -25,30 +27,90 @@
       (seq warnings) (assoc :raw-warnings warnings)
       (debug/dump-configured?) (dump :raw-directives))))
 
-(defn- run-raw-plugins
-  [m]
-  (let [{:keys [directives errors]}
-          (plugins/run-xf (:raw-directives m) (:plugins m) :raw-xf)]
-    (cond-> (assoc m :raw-xf-directives (type/directives directives))
-      (debug/dump-configured?) (dump :raw-xf-directives)
-      (seq errors) (assoc :raw-xf-errors errors))))
+(defn- run-plugins
+  "Run plugins, kind being :raw or :booked"
+  [m kind]
+  (let [kind-name (name kind)
+        key (into {}
+                  (map (fn [k] [k (keyword (str kind-name "-" (name k)))])
+                    [:xf :directives :xf-directives :xf-errors]))
+        create-synthetic-spans-if-required
+          (if (= kind :raw) #(synthetic-spans/create % (:pod m)) identity)]
+    (try (let [{:keys [directives errors]} (plugins/run-plugins-of-kind
+                                             (get m (:directives key))
+                                             (:plugins m)
+                                             (:xf key))]
+           (cond-> (assoc m
+                     (:xf-directives key) (type/directives
+                                            (create-synthetic-spans-if-required
+                                              directives)))
+             (debug/dump-configured?) (dump (:directives key))
+             (seq errors) (assoc (:xf-errors key) errors)))
+         (catch Exception e
+           (binding [*err* *out*]
+             (println "Exception in"
+                      kind-name
+                      "plugin, all"
+                      kind-name
+                      "plugins ignored")
+             (.printStackTrace e))
+           m))))
+
+(defn- resolve-idx
+  [[dct-idx pst-idx] directives]
+  (let [dct (get directives dct-idx)
+        pst (get (:postings dct) pst-idx)]
+    (cond-> {:kind (if pst "posting" (name (:dct dct))),
+             :span (or (:span pst) (:span dct))}
+      pst (assoc :context ["txn" (:span dct)]))))
+
+(defn- resolve-related
+  [directives]
+  (fn [idx]
+    (let [{:keys [kind span]} (resolve-idx idx directives)] [kind span])))
+
+(defn- resolve-indexed-report
+  [report directives]
+  (let [{:keys [kind span context]} (resolve-idx (:idx report) directives)]
+    (cond-> (assoc (select-keys report [:reason :annotation])
+              :message (str "invalid " kind)
+              :span span)
+      context (assoc :context context)
+      (:related report) (assoc :related
+                          (mapv (resolve-related directives)
+                            (:related report))))))
+
+(defn- resolve-indexed-reports
+  [reports directives]
+  (map #(resolve-indexed-report % directives) reports))
 
 (defn- book-raw-directives
-  "Book the raw-xf directives if any, otherwise use the raw plugins as parsed."
+  "Book the raw-xf directives if any, otherwise use the raw directives as parsed."
   [m]
-  (let [{:keys [directives warnings]} (pod/book (:pod m)
-                                                (:raw-xf-directives m))]
-    (cond-> (assoc m :booked-directives (type/directives directives))
-      (debug/dump-configured?) (dump :booked-directives)
-      (seq warnings) (assoc :booked-warnings warnings))))
-
-(defn- run-booked-plugins
-  [m]
-  (let [{:keys [directives errors]}
-          (plugins/run-xf (:booked-directives m) (:plugins m) :booked-xf)]
-    (cond-> (assoc m :booked-xf-directives (type/directives directives))
-      (debug/dump-configured?) (dump :booked-xf-directives)
-      (seq errors) (assoc :booked-xf-errors errors))))
+  (binding [*err* *out*]
+    (try (let [{:keys [ok err]} (pod/book (:pod m) (:raw-xf-directives m))]
+           (if ok
+             (let [{:keys [directives warnings]} ok]
+               (cond-> (assoc m :booked-directives (type/directives directives))
+                 (debug/dump-configured?) (dump :booked-directives)
+                 (seq warnings) (assoc :booked-warnings warnings)))
+             (let [{:keys [spanned-reports indexed-reports message]} err
+                   raw-directives (or (:raw-xf-directives m)
+                                      (:raw-directives m))
+                   resolved-reports (or spanned-reports
+                                        (and indexed-reports
+                                             (resolve-indexed-reports
+                                               indexed-reports
+                                               raw-directives)))
+                   message (if resolved-reports
+                             (pod/format-errors (:pod m) resolved-reports)
+                             message)]
+               (println "Booking failed\n" message)
+               (assoc-in m [:errors :booking] err))))
+         (catch Exception e
+           (println "Booking failed")
+           (exception/print-exception e)
+           m))))
 
 (defn load-beanfile
   [path]
@@ -58,11 +120,11 @@
         resolved-plugins (plugins/resolve-symbols plugins options)]
     (cond-> {:path path, :pod pod, :plugins resolved-plugins, :options options}
       true (get-raw-directives)
-      (plugins/has-specified-plugins? resolved-plugins :raw-xf)
-        (run-raw-plugins)
+      (plugins/has-specified-plugins? resolved-plugins :raw-xf) (run-plugins
+                                                                  :raw)
       true (book-raw-directives)
-      (plugins/has-specified-plugins? resolved-plugins :booked-xf)
-        (run-booked-plugins)
+      (plugins/has-specified-plugins? resolved-plugins :booked-xf) (run-plugins
+                                                                     :booked)
       true (as-> m (assoc m
                      :directives (or (:booked-xf-directives m)
                                      (:booked-directives m)))