+++ clj/src/limabean/core/format.clj
@@ -80,18 +80,49 @@
                   (when-let [total (:total price)] (print " #" (str total))))
                 (when-let [cur (:cur price)] (print "" cur))))
 
+(defn- print-posting
+  [pst]
+  (print " ")
+  (when-let [flag (:flag pst)] (print "" flag))
+  (print "" (:acc pst))
+  (when-let [units (:units pst)] (print "" (str units)))
+  (when-let [cur (:cur pst)] (print "" cur))
+  (when-let [cost (or (:cost pst) (:cost-spec pst))]
+    (print "" (cost->str cost)))
+  (when-let [price (or (:price pst) (:price-spec pst))]
+    (print "" (price->str price)))
+  (println)
+  (print-tags-and-links-on-separate-lines "  " (:tags pst) (:links pst))
+  (when-let [metadata (:metadata pst)]
+    (print-meta-key-values-on-separate-lines "  " metadata)))
+
+(defn- posting->str [pst] (with-out-str (print-posting pst)))
+
+(defn- print-dct-common-header-fields
+  [dct]
+  (print (str (:date dct)) (dct-type dct)))
+
+(defn- print-txn-specific-header-fields
+  [txn]
+  (let [payee (double-quote (:payee txn))
+        narration (double-quote (:narration txn))
+        empty (double-quote "")]
+    (cond (and payee narration) (print "" payee narration)
+          payee (print "" payee empty)
+          narration (print "" narration))))
+
+(defn- txn-header->str
+  [txn]
+  (with-out-str (print-dct-common-header-fields txn)
+                (print-txn-specific-header-fields txn)))
+
 (defn directive->str
   "Convert directive to string"
   [dct]
   (with-out-str
-    (print (str (:date dct)) (dct-type dct))
+    (print-dct-common-header-fields dct)
     (case (:dct dct)
-      :txn (let [payee (double-quote (:payee dct))
-                 narration (double-quote (:narration dct))
-                 empty (double-quote "")]
-             (cond (and payee narration) (print "" payee narration)
-                   payee (print "" payee empty)
-                   narration (print "" narration)))
+      :txn (print-txn-specific-header-fields dct)
       :price (let [price (:price dct)]
                (print "" (:cur dct) (str (:per-unit price)) (:cur price)))
       :balance (do (print "" (:acc dct) (str (:units dct)) (:cur dct))
@@ -122,19 +153,31 @@
     (when-let [metadata (:metadata dct)]
       (print-meta-key-values-on-separate-lines "  " metadata))
     (when (= (:dct dct) :txn)
-      (doseq [posting (:postings dct)]
-        (print " ")
-        (when-let [flag (:flag posting)] (print "" flag))
-        (print "" (:acc posting))
-        (when-let [units (:units posting)] (print "" (str units)))
-        (when-let [cur (:cur posting)] (print "" cur))
-        (when-let [cost (or (:cost posting) (:cost-spec posting))]
-          (print "" (cost->str cost)))
-        (when-let [price (or (:price posting) (:price-spec posting))]
-          (print "" (price->str price)))
-        (println)
-        (print-tags-and-links-on-separate-lines "  "
-                                                (:tags posting)
-                                                (:links posting))
-        (when-let [metadata (:metadata posting)]
-          (print-meta-key-values-on-separate-lines "  " metadata))))))
+      (doseq [pst (:postings dct)] (print-posting pst)))))
+
+(defn elements-xf
+  "Transducer to produce elements from directives, to facilitate building synthetic spans.
+
+  Except for transactions, each directive results in a single string, its human-readable representation.
+
+  Transactions result in strings for the header and each posting separately.
+
+  `element-builder` is a function taking the directive and string, and returning the element
+  "
+  [element-builder]
+  (fn [rf]
+    (fn
+      ;; init
+      ([] (rf))
+      ;; completion
+      ([result] (rf result))
+      ;; step
+      ([result dct]
+       (if (= (:dct dct) :txn)
+         ;; transaction header line and postings separately
+         (let [result' (rf result (element-builder dct (txn-header->str dct)))]
+           (reduce rf
+             result'
+             (map #(element-builder dct (posting->str %)) (:postings dct))))
+         ;; otherwise emit the whole formatted directive
+         (rf result (element-builder dct (directive->str dct))))))))